+++ src/components/cardFormFields/CardFormFields.stories.tsx
@@ -1,21 +1,21 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import CardInfo from "./CardInfo";
+import CardFormFields from "./CardFormFields";
 import { useCardForm } from "../useCardForm";
 import { useCardStep } from "./useCardStep";
 
-const meta: Meta<typeof CardInfo> = {
-  title: "Components/CardInfo",
-  component: CardInfo,
+const meta: Meta<typeof CardFormFields> = {
+  title: "Components/CardFormFields",
+  component: CardFormFields,
 };
 
 export default meta;
 
-type Story = StoryObj<typeof CardInfo>;
+type Story = StoryObj<typeof CardFormFields>;
 
 export const Default: Story = {
   render: () => {
     const cardForm = useCardForm();
     const step = useCardStep(cardForm);
-    return <CardInfo cardForm={cardForm} step={step} />;
+    return <CardFormFields cardForm={cardForm} step={step} />;
   },
 };
